+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Hello App Display Name</string>
+    <string name="app_name">Home</string>
 </resources>
